[build] don't install anything globally outside virtualenv
I don't want that my global python configuration is modified when I build Noto fonts.
It's better to avoid any `sudo` and even `pip install --user`, and let the user
decide what to do.

If pip is missing, we just fail, suggest a course of action and direct the user to the
pip documentation.

If pip is installed but virtualenv is not, we can install virtualenv in our
working directory with --target option, and use it to bootstrap our virtual environment.

Also, before installing fontmake, it's a good idea to make sure the submodule is
up-to-date.